Lab 1
This is the standard palindrome problem with a slight variation.

You are provided with a string consisting of alpha-numeric characters. Return 1 if the string is palindromic, 0 otherwise.

A string is palindromic if it is the same as its reverse. Here, we consider lowercase letters to be the same as the uppercase letters, i.e., 'a' and 'A' are considered to be equal.

Note:

Input and Output are handled automatically, you just need to implement the definition of function isPalindrome(s).
Do not change any predefined code. Doing so would lead to deduction of marks.
Input Format

You are given a method isPalindrome(s). s is a string containing only alpha-numeric characters without any spaces. Alphabets can be lowercase or uppercase.

Constraints

1 ≤ N ≤ 10, where N is the length of the string
Output Format

Return an integer value, 1 or 0. Return 1 if the string is palindromic, and 0 otherwise.

Sample Input 0

A11a
Sample Output 0

1
Explanation 0

Reversing the string, we get a11A. As we consider a and A to be the same, the string is a palindrome.