Lab 1 Q2
You are given a list of Integers say A = {a1, a2, a3, ... an} with n elements.

You need to return the number of ordered pair a[i] and a[j] where i != j and a[i] + a[j] = c.

Ordered pair means, for some i and j, (a[i], a[j]) and (a[j], a[i]) are different.

eg (a[0], a[1]) and (a[1], a[0]) are 2 pairs.

Note:

Input and Output are handled automatically, you just need to implement the definition of function numPairs(l, n, c):.
Do not change any predefined code. Doing so would lead to deduction of marks.
Input Format

1st Line Contains Integer n
2nd Line Contains Integer c
3rd Line Contains N space separated integers, a1 a2 a3 ... an
Input have been handled. You need to implement your logic under **def numPairs(l, n, c): **only.
Constraints

1 ≤ n ≤ 10
-100 ≤ c ≤ 100
-100 ≤ Ai ≤ 100
Output Format

Return a single integer, the number of such pairs whose sum equals c.
Sample Input 0

5 
4
2 2 1 3 4
Sample Output 0

4
Explanation 0

Here, n = 5, c = 4

We need to find all the pairs whose sum = 4,
For index i = 0 and j = 1, a[i] + a[j] = 2 + 2 = 4
For index i = 1 and j = 0, a[i] + a[j] = 2 + 2 = 4
For index i = 2 and j = 3, a[i] + a[j] = 1 + 3 = 4
For index i = 3 and j = 1, a[i] + a[j] = 3 + 1 = 4

No other such pairs are possible, therefore ans = 4